backend/requirements.txt
asgiref==3.8.1
channels==4.2.2
Django==5.2.1
djangorestframework==3.16.0
sqlparse==0.5.3
tzdata==2025.2
